+++ metadata.rb
@@ -1,10 +1,10 @@
-name 'cassandra'
+name 'cassandra-dse'
 maintainer 'Michael S. Klishin'
 maintainer_email 'michael@clojurewerkz.org'
 license 'Apache 2.0'
 description 'Installs/configures Apache Cassandra'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version '2.10.0'
+version '3.2.0'
 depends 'java'
 depends 'ulimit'
 depends 'apt'