+++ recipes/opscenter_server.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: cassandra
+# Cookbook Name:: cassandra-dse
 # Recipe:: opscenter_server
 #
 # Copyright 2011-2012, Michael S Klishin & Travis CI Development Team
@@ -18,7 +18,7 @@
 #
 
 include_recipe 'java' if node['cassandra']['install_java']
-include_recipe 'cassandra::repositories'
+include_recipe 'cassandra-dse::repositories'
 
 ops = node['cassandra']['opscenter']
 ops_server = ops['server']