+++ recipes/tarball.rb
@@ -1,7 +1,7 @@
 #
-# Cookbook Name:: cassandra
+# Cookbook Name:: cassandra-dse
 # Recipe:: tarball
-# Copyright 2012, Michael S. Klishin <michaelklishin@me.com>
+# Copyright 2012-2015, Michael S. Klishin and Travis CI Development Team <michael@clojurewerkz.org>
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -42,7 +42,7 @@ include_recipe 'java' if node['cassandra']['install_java']
 Chef::Application.fatal!("attribute node['cassandra']['cluster_name'] not defined") unless node['cassandra']['cluster_name']
 
 # 2. Manage C* Service User
-include_recipe 'cassandra::user'
+include_recipe 'cassandra-dse::user'
 
 # 3. Download the tarball to /tmp
 require 'tmpdir'