+++ spec/default_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe 'cassandra::default' do
+describe 'cassandra-dse::default' do
 
   shared_examples_for 'cassandra' do
 
@@ -103,8 +103,8 @@ describe 'cassandra::default' do
       expect(chef_run).to_not create_link('/usr/share/cassandra/lib/jna.jar').with(to: '/usr/share/java/jna.jar')
     end
 
-    it 'Creates symlink between /usr/share/cassandra/lib/jamm-0.2.6.jar and /usr/share/java/jamm-0.2.6.jar' do
-      expect(chef_run).to create_link('/usr/share/cassandra/lib/jamm-0.2.6.jar').with(to: '/usr/share/java/jamm-0.2.6.jar')
+    it 'Creates symlink between /usr/share/cassandra/lib/jamm-0.2.8.jar and /usr/share/java/jamm-0.2.8.jar' do
+      expect(chef_run).to create_link('/usr/share/cassandra/lib/jamm-0.2.8.jar').with(to: '/usr/share/java/jamm-0.2.8.jar')
     end
 
     it 'creates /etc/cassandra/conf/logback.xml' do